Better support DerivedData; fix minor regressions roundtripping to N42-2012; code cleanup. Tried a few more vendors N42 files with DerivedData and made some tweaks. Fixed issue, or rather duct taped around, reading and writing DetectorKindCode from N42-2012 files. Fixed regression where one N42-2006 file had some discrepancy when serialized to a N42-2012 file and then read back in; the sample number and Measurement ordering would differ.  The solution implemented isn't great, and should probably be revisited as it impacts performance of writing large N42-2012 files, and it alters how they are written for some otherwise not affected input files. Did some minor code cleanup, replacing manual rapidxml accessing code to use of macros. Fixed issue reading GPS coordinates from SPE files. Made some notes for some things that should be changed in the future.

+++ src/SpecFile_cnf.cpp
@@ -158,11 +158,11 @@ namespace
     {
         std::array< byte, sizeof(int64_t) > bytes = { 0x00 };
         //duration in usec is larger than a int64: covert to years
-        if (duration * 10000000 > INT64_MAX)
+        if ( (static_cast<double>(duration) * 10000000.0) > static_cast<double>(INT64_MAX) )
         {
             double t_duration = duration / 31557600;
             //duration in years is larger than an int32, divide by a million years
-            if (duration / 31557600 > INT32_MAX)
+            if ( (duration / 31557600.0) > static_cast<double>(INT32_MAX) )
             {
                 int32_t y_duration = static_cast<int32_t>(t_duration / 1e6);
                 const auto y_bytes = to_bytes(y_duration);
